requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.6
gunicorn==20.0.4
pandas==0.23.3
numpy==1.15.1
google-cloud-bigquery==1.9.0
beautifulsoup4==4.8.1
google-cloud-storage==1.13.2
requests==2.22.0
scipy==1.2.1
alpaca-trade-api==0.41
pyportfolioopt==0.4.3
scikit-learn==0.21.3
alpha_vantage-2.1.2
